feat/Client : refactoring update subcategories code for accountant
